Complete day 9 part 2

+++ 2018/09/Marble.java
@@ -1,10 +1,10 @@
 public class Marble {
 
-    public int value;
+    public Long value;
     public Marble next;
     public Marble prev;
 
-    public Marble(int value, Marble next, Marble prev) {
+    public Marble(Long value, Marble next, Marble prev) {
         this.value = value;
         this.next = next;
         this.prev = prev;
